refactor: rename src/logging -> src/log-format to fix collision with src/logger
src/logger (structured JSONL Logger facility) and src/logging (human-facing
presentation/formatter that consumes LogEntry and renders run summaries) are
distinct, cleanly-layered subsystems — not duplicates. Their near-identical
names are a readability footgun. Rename the smaller presentation layer to
src/log-format to remove the ambiguity.

- git mv src/logging -> src/log-format, test/unit/logging -> test/unit/log-format
- Update 5 import sites (headless-formatter, logger, logs-formatter, test)
- Clarify the barrel header comment to disambiguate from src/logger
- No behavior change
