--- src/logging/index.ts
@@ -1,22 +1,0 @@
-/**
- * Logging formatter module
- *
- * Provides human-friendly log formatting with multiple verbosity levels
- */
-
-export {
-  formatLogEntry,
-  formatRunSummary,
-  formatTimestamp,
-  formatDuration,
-  formatCost,
-  type FormattedEntry,
-} from "./formatter.js";
-export {
-  EMOJI,
-  type VerbosityMode,
-  type FormatterOptions,
-  type RunSummary,
-  type StoryStartData,
-  type StageResultData,
-} from "./types.js";
